refactor: split dynamic-table tests out of main_test.go
main_test.go hit qlty's per-file complexity threshold (50). Move
all TestDynamicTableHandler_* and TestListDynamicTablesHandler_*
cases plus the withDynamicTables helper into dynamic_tables_test.go.
Also drop imports main_test.go no longer needs (os, path/filepath,
dynamictablefilter). qlty smells now returns zero.
